+++ banyand/stream/introducer.go
@@ -221,6 +221,7 @@ func (tst *tsTable) introduceMemPart(nextIntroduction *introduction, epoch uint6
 	}
 
 	next := nextIntroduction.memPart
+	tst.addPendingDataCount(-int64(next.mp.partMetadata.TotalCount))
 	nextSnp := cur.copyAllTo(epoch)
 	nextSnp.parts = append(nextSnp.parts, next)
 	nextSnp.creator = snapshotCreatorMemPart